Refactor pack registry code out of `config-utils`
